% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DML_CKLS.R
\name{est.CKLS.DML}
\alias{est.CKLS.DML}
\title{ML estimation for the CKLS model (Euler method)}
\usage{
est.CKLS.DML(X, Delta = deltat(X), par = NULL)
}
\arguments{
\item{X}{a numeric vector, the sample path of the SDE.}

\item{Delta}{a single numeric, the time step between two consecutive observations.}

\item{par}{a numeric vector with dimension four indicating initial values of the
parameters. Defaults to NULL, fits a linear model using generalized least squares
with AR1 correlation and a power variance heteroscedasticity structure.}
}
\value{
A list containing a matrix with the estimated coefficients and the
associated standard errors.
}
\description{
Parametric estimation for the CKLS model using maximum likelihood and
the discretized version of the model, obtained with the Euler-Maruyama
method. The parametric form of the CKLS model used here is given by
\deqn{dX_t = (\alpha - \kappa X_t)dt + \sigma X_t^\gamma dW_t.}
}
\examples{
x <- rCKLS(360, 1/12, 0.09, 0.08, 0.9, 1.2, 1.5)
est.CKLS.DML(x)
}
