% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM.R
\name{est.CKLS.GMM}
\alias{est.CKLS.GMM}
\title{Generalized method of moments estimator for the CKLS model}
\usage{
est.CKLS.GMM(X, Delta = deltat(X), par = NULL, maxiter = 25)
}
\arguments{
\item{X}{a numeric vector, the sample path of the SDE.}

\item{Delta}{a single numeric, the time step between two consecutive observations.}

\item{par}{a numeric vector with dimension four indicating initial values of the
parameters. Defaults to NULL, fits a linear model using generalized least squares
with AR1 correlation and a power variance heteroscedasticity structure.}

\item{maxiter}{an integer, the maximum number of iterations.}
}
\value{
A list containing a matrix with the estimated coefficients and the
associated standard errors.
}
\description{
Parametric estimation for the CKLS model using the Generalized Method
of Moments. The parametric form of the CKLS model used here is given by
\deqn{dX_t = (\alpha - \kappa X_t)dt + \sigma X_t^\gamma dW_t.}
}
\examples{
x <- rCKLS(360, 1/12, 0.09, 0.08, 0.9, 1.2, 1.5)
est.CKLS.GMM(x)

}
\references{
Hansen, L. P. (1982). Large sample properties of generalized method of moments
estimators. Econometrica, pages 1029–1054.

Chan, K. C., Karolyi, G. A., Longstaff, F. A., and Sanders, A. B. (1992).
An empirical comparison of alternative models of the short-term interest rate.
The journal of finance, 47(3):1209–1227.
}
