% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KF_CKLS.R
\name{est.CKLS.KF}
\alias{est.CKLS.KF}
\title{ML estimation for the CKLS model (Kalman filter)}
\usage{
est.CKLS.KF(X, Delta = deltat(X), par = NULL, mu0 = 0, Sigma0 = 1)
}
\arguments{
\item{X}{a numeric vector, the sample path of the SDE.}

\item{Delta}{a single numeric, the time step between two consecutive observations.}

\item{par}{a numeric vector with dimension four indicating initial values of the
parameters. Defaults to NULL, fits a linear model using generalized least squares
with AR1 correlation and a power variance heteroscedasticity structure.}

\item{mu0}{a single numeric, the initial mean. Defaults to zero.}

\item{Sigma0}{a single numeric, the initial variance. Defaults to one.}
}
\value{
A list containing a matrix with the estimated coefficients and the
associated standard errors.
}
\description{
Parametric estimation for the CKLS model using the Kalman filter algorithm, where
the discretized version of the model is obtained with the Euler-Maruyama method.
The parametric form of the CKLS model used here is given by
\deqn{dX_t = (\alpha - \kappa X_t)dt + \sigma X_t^\gamma dW_t.}
}
\examples{
x <- rCKLS(360, 1/12, 0.09, 0.08, 0.9, 1.2, 1.5)
est.CKLS.KF(x)
}
