% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL_CKLS.R
\name{est.CKLS.LL}
\alias{est.CKLS.LL}
\title{ML estimation for the CKLS model (local linearization)}
\usage{
est.CKLS.LL(X, Delta = deltat(X), par = NULL)
}
\arguments{
\item{X}{a numeric vector, the sample path of the SDE.}

\item{Delta}{a single numeric, the time step between two consecutive observations.}

\item{par}{a numeric vector with dimension four indicating initial values of the
parameters. Defaults to NULL, fits a linear model using generalized least squares
with AR1 correlation and a power variance heteroscedasticity structure.}
}
\value{
A list containing a matrix with the estimated coefficients and the
associated standard errors.
}
\description{
Parametric estimation for the CKLS model using maximum likelihood and
the discretized version of the model, obtained with the local linearization
method. The parametric form of the CKLS model used here is given by
\deqn{dX_t = (\alpha - \kappa X_t)dt + \sigma X_t^\gamma dW_t.}
}
\examples{
x <- rCKLS(360, 1/12, 0.09, 0.08, 0.9, 1.2, 1.5)
est.CKLS.LL(x)

}
\references{
Ozaki, T. (1992). A bridge between nonlinear time series models and
nonlinear stochastic dynamical systems: a local linearization approach.
Statistica Sinica, pages 113–135.

Shoji, I. and Ozaki, T. (1998). Estimation for nonlinear stochastic
differential equations by a local linearization method. Stochastic
Analysis and Applications, 16(4):733–752.
}
