% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{est.CKLS.MCMC}
\alias{est.CKLS.MCMC}
\title{MCMC estimation for the CKLS model}
\usage{
est.CKLS.MCMC(X, Delta = deltat(X), par = NULL, niter = 4000, burn_in = 1000)
}
\arguments{
\item{X}{a numeric vector, the sample path of the SDE.}

\item{Delta}{a single numeric, the time step between two consecutive observations.}

\item{par}{a numeric vector with dimension four indicating initial values of the
parameters. Defaults to NULL, fits a linear model using generalized least squares
with AR1 correlation and a power variance heteroscedasticity structure.}

\item{niter}{an integer, number of iterations.}

\item{burn_in}{an integer indicating the number of initial iterations to be discarded.}
}
\value{
A list containing a matrix with the estimated coefficients and the
associated standard errors.
}
\description{
Parametric estimation for the CKLS model using Markov Chain Monte Carlo and
involving data augmentation, as proposed in Elerian et al. (2001) and Eraker (2001).
The parametric form of the CKLS model used here is given by
\deqn{dX_t = (\alpha - \kappa X_t)dt + \sigma X_t^\gamma dW_t.}
}
\examples{
\donttest{
set.seed(987)
x <- rCKLS(480, 1/12, 0.09, 0.08, 0.9, 1.2, 1.5)
est.CKLS.MCMC(x)
}

}
\references{
Elerian, O., Chib, S., and Shephard, N. (2001). Likelihood inference for discretely
observed nonlinear diffusions. Econometrica, 69(4):959–993.

Eraker, B. (2001). MCMC analysis of diffusion models with application to finance.
Journal of Business & Economic Statistics, 19(2):177–191.
}
