% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM.R
\name{est.VAS.GMM}
\alias{est.VAS.GMM}
\title{Generalized method of moments estimator for the Vasicek model}
\usage{
est.VAS.GMM(X, Delta = deltat(X), par = NULL, maxiter = 25)
}
\arguments{
\item{X}{a numeric vector, the sample path of the SDE.}

\item{Delta}{a single numeric, the time step between two consecutive observations.}

\item{par}{a numeric vector with dimension three indicating initial values of the
parameters. Defaults to NULL, fits a linear model as an initial guess.}

\item{maxiter}{an integer, the maximum number of iterations.}
}
\value{
A list containing a matrix with the estimated coefficients and the
associated standard errors.
}
\description{
Parametric estimation for the Vasicek model using the Generalized Method
of Moments. The parametric form of the Vasicek model used here is given by
\deqn{dX_t = (\alpha - \kappa X_t)dt + \sigma dW_t.}
}
\examples{
x <- rVAS(360, 1/12, 0, 0.08, 0.9, 0.1)
est.VAS.GMM(x)

}
\references{
Hansen, L. P. (1982). Large sample properties of generalized method of moments
estimators. Econometrica, pages 1029–1054.
}
