% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KF_OU.R
\name{est.VAS.KF}
\alias{est.VAS.KF}
\title{ML estimation for the Vasicek model (Kalman filter)}
\usage{
est.VAS.KF(X, Delta = deltat(X), par = NULL, mu0 = 0, Sigma0 = 1)
}
\arguments{
\item{X}{a numeric vector, the sample path of the SDE.}

\item{Delta}{a single numeric, the time step between two consecutive observations.}

\item{par}{a numeric vector with dimension three indicating initial values of the
parameters. Defaults to NULL, fits a linear model as an initial guess.}

\item{mu0}{a single numeric, the initial mean. Defaults to zero.}

\item{Sigma0}{a single numeric, the initial variance. Defaults to one.}
}
\value{
A list containing a matrix with the estimated coefficients and the
associated standard errors.
}
\description{
Parametric estimation for the Vasicek model using the Kalman filter algorithm, where
the discretized version of the model is obtained with the Euler-Maruyama method.
The parametric form of the Vasicek model used here is given by
\deqn{dX_t = (\alpha - \kappa X_t)dt + \sigma dW_t.}
}
\examples{
x <- rVAS(360, 1/12, 0, 0.08, 0.9, 0.1)
est.VAS.KF(x)
}
