% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL_OU.R
\name{est.VAS.LL}
\alias{est.VAS.LL}
\title{ML estimation for the Vasicek model (local linearization)}
\usage{
est.VAS.LL(X, Delta = deltat(X), par = NULL)
}
\arguments{
\item{X}{a numeric vector, the sample path of the SDE.}

\item{Delta}{a single numeric, the time step between two consecutive observations.}

\item{par}{a numeric vector with dimension three indicating initial values of the
parameters. Defaults to NULL, fits a linear model as an initial guess.}
}
\value{
A list containing a matrix with the estimated coefficients and the
associated standard errors.
}
\description{
Parametric estimation for the Vasicek model using maximum likelihood and
the discretized version of the model, obtained with the local linearization
method. The parametric form of the Vasicek model used here is given by
\deqn{dX_t = (\alpha - \kappa X_t)dt + \sigma dW_t.}
}
\examples{
x <- rVAS(360, 1/12, 0, 0.08, 0.9, 0.1)
est.VAS.LL(x)

}
\references{
Ozaki, T. (1992). A bridge between nonlinear time series models and
nonlinear stochastic dynamical systems: a local linearization approach.
Statistica Sinica, pages 113–135.

Shoji, I. and Ozaki, T. (1998). Estimation for nonlinear stochastic
differential equations by a local linearization method. Stochastic
Analysis and Applications, 16(4):733–752.
}
