% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCKLS.R
\name{rCKLS}
\alias{rCKLS}
\title{Simulation of the CKLS model}
\usage{
rCKLS(n, Delta, X0, alpha, kappa, sigma, gamma)
}
\arguments{
\item{n}{an integer indicating the sample size.}

\item{Delta}{a single numeric, the time step between two consecutive observations.}

\item{X0}{a single numeric, initial value of the process.}

\item{alpha}{a single numeric, drift parameter.}

\item{kappa}{a single numeric, rate of mean reversion.}

\item{sigma}{a single numeric, volatility parameter.}

\item{gamma}{a single numeric, proportional volatility exponent.}
}
\value{
A \code{ts} object, a numeric vector with the sample path of the SDE.
}
\description{
\code{rCKLS} simulates a single path of the CKLS process using the Milstein scheme.
The parametric form of the CKLS model used here is given by
\deqn{dX_t = (\alpha - \kappa X_t)dt + \sigma X_t^\gamma dW_t.}
}
\examples{
x <- rCKLS(360, Delta = 1/12, X0 = 0.09, alpha = 0.08, kappa = 0.9, sigma = 1.2, gamma = 1.5)
plot(x)

}
\references{
Chan, K. C., Karolyi, G. A., Longstaff, F. A., and Sanders, A. B. (1992).
An empirical comparison of alternative models of the short-term interest rate.
The journal of finance, 47(3):1209–1227.
}
