% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simVAS.R
\name{rVAS}
\alias{rVAS}
\title{Simulation of the Vasicek model}
\usage{
rVAS(n, Delta, X0, alpha, kappa, sigma)
}
\arguments{
\item{n}{an integer indicating the sample size.}

\item{Delta}{a single numeric, the time step between two consecutive observations.}

\item{X0}{a single numeric, initial value of the process.}

\item{alpha}{a single numeric, drift parameter.}

\item{kappa}{a single numeric, rate of mean reversion.}

\item{sigma}{a single numeric, volatility parameter.}
}
\value{
A \code{ts} object, a numeric vector with the sample path of the SDE.
}
\description{
\code{rVAS} simulates a single path of the Vasicek process using its known conditional
distribution. The parametric form of the Vasicek model used here is given by
\deqn{dX_t = (\alpha - \kappa X_t)dt + \sigma dW_t.}
}
\examples{
x <- rVAS(360, Delta = 1/12, X0 = 0.09, alpha = 0.08, kappa = 0.9, sigma = 0.1)
plot(x)

}
